--- Assets/Externals/HQ Shipping Container (Modular)/HQ Shipping Container (Modular) Mobile Package/HQ Shipping Container (Modular) Mobile Package.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a467e57ab0f8f39468f833310cdae5e6
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
