--- Assets/Externals/LowPolyFPSLite/LowPolyFPSLite_2021.3.0_URP_v1.0.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 718863dcff53b784aa0ffc114381cdd7
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
