--- Assets/Scripts/Targets/HiddenTargetPattern.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2e35dff90ccc5e34e9b8851134058689
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
