--- Assets/Scripts/Targets/TargetMovement.cs
@@ -1,37 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class TargetMovement : MonoBehaviour
-{
-    public GameObject targetPrefab;
-    public float moveSpeed = 4.0f; //움직임 속도
-    public float moveDistance = 15.0f; //좌우로 움직일 거리
-
-    private Vector3 initialPosition;
-    private bool movingRight = true;
-    // Start is called before the first frame update
-    void Start()
-    {
-        initialPosition = transform.position;
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        //좌우로 움직이는 로직
-        if(movingRight)
-        {
-            transform.Translate(Vector3.right * moveSpeed * Time.deltaTime);
-        }
-        else
-        {
-            transform.Translate(Vector3.left * moveSpeed * Time.deltaTime);
-        }
-        //일정 거리를 이동하면 방향을 바꿈
-        if (Vector3.Distance(initialPosition, transform.position) >= moveDistance)
-        {
-            movingRight = !movingRight;
-        }
-    }
-}
