--- Assets/Scripts/Targets/TargetMovement.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: ac4bfd23b7e2808448bdcbf593f06494
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences:
-  - targetPrefab: {fileID: 1336443488165784, guid: c7e9f74480d2343ef87c5c15624833d2,
-      type: 3}
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
