+++ Assets/Scripts/Targets/TargetMovingSpawn.cs
@@ -2,36 +2,48 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class MovingTargetSpawner : MonoBehaviour
+public class TargetMovingSpawn : MonoBehaviour
 {
-    public GameObject targetPrefab; // 움직일 타겟 프리팹
-    public Transform spawnPoint; // 타겟을 소환할 위치
-    public float moveSpeed = 4.0f; // 타겟의 이동 속도
-
-    private GameObject currentTarget; // 현재 타겟
-
+    public GameObject targetPrefab; // 타겟 프리팹
+    public Transform targetPosition; // 타겟을 생성할 위치
+    public float moveSpeed = 4.0f; //움직임 속도
+    public float moveDistance = 10.0f; //좌우로 움직일 거리
+    private Vector3 initialPosition;
+    private bool movingRight = true;
     void Start()
     {
-        SpawnMovingTarget();
+        // 시작할 때 타겟을 생성하고 위치 설정
+        TargetSpawn(targetPosition.position);
+        initialPosition = transform.position;
     }
 
-    void SpawnMovingTarget()
+    void TargetSpawn(Vector3 position)
     {
-        if (currentTarget != null)
-        {
-            // 이전 타겟이 있으면 제거
-            Destroy(currentTarget);
-        }
-
         // 타겟을 생성하고 위치 설정
-        currentTarget = Instantiate(targetPrefab, spawnPoint.position, Quaternion.identity);
+        GameObject newTarget = Instantiate(targetPrefab, position, Quaternion.identity);
 
-        // 타겟에 움직임을 제어하는 스크립트를 추가
-        MovingTargetScript movingScript = currentTarget.GetComponent<MovingTargetScript>();
-        if (movingScript != null)
+        // 여기에서 타겟을 초기화하거나 추가적인 작업을 수행할 수 있습니다.
+
+        // 생성한 타겟을 원하는 위치로 이동하거나 조작할 수 있습니다.
+
+    }
+   
+
+    void Update()
+    {
+        //좌우로 움직이는 로직
+        if (movingRight)
         {
-            // 움직임 속도 설정
-            movingScript.SetMoveSpeed(moveSpeed);
+            transform.Translate(Vector3.right * moveSpeed * Time.deltaTime);
+        }
+        else
+        {
+            transform.Translate(Vector3.left * moveSpeed * Time.deltaTime);
+        }
+        //일정 거리를 이동하면 방향을 바꿈
+        if (Vector3.Distance(initialPosition, transform.position) >= moveDistance)
+        {
+            movingRight = !movingRight;
         }
     }
 }
