--- Assets/Scripts/Targets/TargetMovingStripts.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fb5d79b13c6c4e84bb7f461450ce3bde
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
